+++ x/exchange/client/cli/tx_setup_test.go
@@ -732,7 +732,7 @@ func TestMakeMsgMarketCommitmentSettle(t *testing.T) {
 		},
 		{
 			name:      "from file",
-			clientCtx: newClientContextWithCodec(t),
+			clientCtx: newClientContext(t),
 			flags:     []string{"--file", filename},
 			expMsg:    fileMsg,
 		},
@@ -742,7 +742,7 @@ func TestMakeMsgMarketCommitmentSettle(t *testing.T) {
 				"--file", filename, "--tag", "new-thang", "--authority",
 				"--outputs", "monroe:87plum",
 			},
-			clientCtx: newClientContextWithCodec(t),
+			clientCtx: newClientContext(t),
 			expMsg: &exchange.MsgMarketCommitmentSettleRequest{
 				Admin:    cli.AuthorityAddr.String(),
 				MarketId: 4,