+++ x/marker/keeper/msg_server.go
@@ -820,3 +820,19 @@ func (k msgServer) SetDenomMetadataProposal(goCtx context.Context, msg *types.Ms
 
 	return &types.MsgSetDenomMetadataProposalResponse{}, nil
 }
+
+// UpdateParams is a governance proposal endpoint for updating the marker module's params.
+func (k msgServer) UpdateParams(goCtx context.Context, msg *types.MsgUpdateParamsRequest) (*types.MsgUpdateParamsResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	if err := k.ValidateAuthority(msg.Authority); err != nil {
+		return nil, err
+	}
+
+	k.SetParams(ctx, msg.Params)
+	if err := ctx.EventManager().EmitTypedEvent(types.NewEventMarkerParamsUpdated(msg.Params.EnableGovernance, msg.Params.GetUnrestrictedDenomRegex(), msg.Params.MaxSupply)); err != nil {
+		return nil, err
+	}
+
+	return &types.MsgUpdateParamsResponse{}, nil
+}